wire: a transmitted bitmask is a namespace of constants, not an enum class
`clang-tidy` found the shape wrong, and it was right rather than pedantic:
`performance-enum-size` on `enum class WireMembershipRoute : std::uint32_t`, whose named
values reach `0x10` and so carry one byte in four.

The tempting fix is to narrow the base type, and it is wrong twice: it caps a SET the wire
has room to grow at eight, and it makes the type's width disagree with the field's. The
right shape was already in the same header and this missed it -- `NodeComponentBit`, the
other transmitted bitmask on this wire, is a namespace of `constexpr std::uint32_t`. A
bitmask is not an enumeration: its values combine, so no variable of the type ever holds
one of them and a `switch` over it means nothing. Every call site keeps its spelling, the
field keeps its width, and there is no enum left to be the wrong size -- and 23 casts to
the type a value already had go with it.

Six optional accesses go through `Unwrap` as well: `bugprone-unchecked-optional-access`
fired on five of them, because the analyser cannot see through Catch2's `REQUIRE`, which
is the whole reason this tree has that helper and a rule saying a bare `*x` is a build
failure. The sixth is the same pattern in a lambda the analyser did not reach; its silence
there is not a licence.

Both are recorded in `.agent/rules/wire-and-protocol.md` with a tripwire in `AGENT.md`,
alongside the other four places a new `0xFC` verb touches -- none of which fails the
BUILD, which is why this branch needed CI to find two of them.

Refs #1471

+++ src/FastCache/Protocol/CompileCacheWire_test.cpp
@@ -2578,11 +2578,11 @@ TEST_CASE("The explain-admission verb occupies the byte it was assigned, in the 
     CHECK(static_cast<std::uint8_t>(WireMembership::Outsider) == 0x01);
     CHECK(static_cast<std::uint8_t>(WireMembership::Member) == 0x02);
     CHECK(static_cast<std::uint8_t>(WireMembership::Forgotten) == 0x03);
-    CHECK(static_cast<std::uint32_t>(WireMembershipRoute::FleetMemberList) == 0x01);
-    CHECK(static_cast<std::uint32_t>(WireMembershipRoute::ClusterMembers) == 0x02);
-    CHECK(static_cast<std::uint32_t>(WireMembershipRoute::ClientTombstone) == 0x04);
-    CHECK(static_cast<std::uint32_t>(WireMembershipRoute::OpenPolicy) == 0x08);
-    CHECK(static_cast<std::uint32_t>(WireMembershipRoute::ProvenKeyHolder) == 0x10);
+    CHECK(WireMembershipRoute::FleetMemberList == 0x01);
+    CHECK(WireMembershipRoute::ClusterMembers == 0x02);
+    CHECK(WireMembershipRoute::ClientTombstone == 0x04);
+    CHECK(WireMembershipRoute::OpenPolicy == 0x08);
+    CHECK(WireMembershipRoute::ProvenKeyHolder == 0x10);
 }
 
 TEST_CASE("An explain-admission request carries exactly one host, and anything else is refused", "[wire][admission]")
@@ -2613,15 +2613,14 @@ TEST_CASE("An admission explanation round-trips every verdict and the whole rout
 {
     // A different route set per verdict, so an encoder that dropped either half cannot
     // agree with all three.
-    for (auto const& sent:
-         { AdmissionExplanationFields { .verdict = WireMembership::Member,
-                                        .decidedBy = static_cast<std::uint32_t>(WireMembershipRoute::FleetMemberList)
-                                                     | static_cast<std::uint32_t>(WireMembershipRoute::ClusterMembers) },
-           AdmissionExplanationFields { .verdict = WireMembership::Forgotten,
-                                        .decidedBy = static_cast<std::uint32_t>(WireMembershipRoute::ClientTombstone) },
-           // The silence: refused, and no route claims authorship. Zero is the READING
-           // here rather than a missing field.
-           AdmissionExplanationFields { .verdict = WireMembership::Outsider, .decidedBy = 0 } })
+    for (auto const& sent: { AdmissionExplanationFields { .verdict = WireMembership::Member,
+                                                          .decidedBy = WireMembershipRoute::FleetMemberList
+                                                                       | WireMembershipRoute::ClusterMembers },
+                             AdmissionExplanationFields { .verdict = WireMembership::Forgotten,
+                                                          .decidedBy = WireMembershipRoute::ClientTombstone },
+                             // The silence: refused, and no route claims authorship. Zero is the READING
+                             // here rather than a missing field.
+                             AdmissionExplanationFields { .verdict = WireMembership::Outsider, .decidedBy = 0 } })
     {
         auto const back = DecodeAdmissionExplanation(EncodeAdmissionExplanation(sent));
         REQUIRE(back.has_value());
@@ -2631,8 +2630,7 @@ TEST_CASE("An admission explanation round-trips every verdict and the whole rout
 
 TEST_CASE("An unknown VERDICT is refused and an unknown ROUTE is kept, which is not one rule twice", "[wire][admission]")
 {
-    auto const routes =
-        WireFields::ToBigEndian<std::uint32_t>(static_cast<std::uint32_t>(WireMembershipRoute::FleetMemberList));
+    auto const routes = WireFields::ToBigEndian<std::uint32_t>(WireMembershipRoute::FleetMemberList);
 
     SECTION("a verdict byte this build cannot name is refused, never read as a refusal nobody authored")
     {
@@ -2652,10 +2650,8 @@ TEST_CASE("An unknown VERDICT is refused and an unknown ROUTE is kept, which is 
         // decided this than did, on a fleet mid-upgrade -- and the reader can tell,
         // because the bit it does not know is still there to count.
         constexpr auto ahead = std::uint32_t { 0x8000'0000 };
-        auto const sent =
-            AdmissionExplanationFields { .verdict = WireMembership::Member,
-                                         .decidedBy =
-                                             static_cast<std::uint32_t>(WireMembershipRoute::FleetMemberList) | ahead };
+        auto const sent = AdmissionExplanationFields { .verdict = WireMembership::Member,
+                                                       .decidedBy = WireMembershipRoute::FleetMemberList | ahead };
         auto const back = DecodeAdmissionExplanation(EncodeAdmissionExplanation(sent));
         REQUIRE(back.has_value());
         CHECK(Unwrap(back).decidedBy == sent.decidedBy);
